databases/work_with_database/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6

django-autoslug~=1.9.8